backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-multipart
passlib
bcrypt
python-jose
requests
APScheduler==3.10.4

